requirements.txt
python-telegram-bot==21.4
python-dotenv==1.0.1
ping3==4.0.8
pytz==2024.1